+++ CHANGES.md
@@ -1,3 +1,12 @@
+# 1.2.0
+
+## API change:
+
+* Changed `Operation#validate`: the signature now is `validate(params, model=self.model, options={}, contract_class)`.
+* Changed `Operation#contract`: the signature now is `contract(model=self.model, options={}, contract_class)`.
+
+Both changes simplify injecting additional dependencies into the contract, or making use of `Composition` contracts. Read more [here](trailblazer.to/gems/operation/contract.html). Note that in 1.2, there are deprecation mechanics to handle old code.
+
 # 1.1.1
 
 * Rename `Operation::Representer::ClassMethods` to `Operation::Representer::DSL` and allow to use `DSL` and `Rendering` without `Deserialization` so you can use two different representers.